requirements.txt
openai
python-dotenv
loguru
nltk
matplotlib
numpy
tiktoken
pyyaml
jinja2